fix issues

+++ _unittests/ut_torch_models/test_llm_model_helper.py
@@ -199,7 +199,7 @@ class TestLlmModelHelper(ExtTestCase):
 
         model, model_inputs = get_phi35_mini_instruct(num_hidden_layers=1)
         model = model
-        with torch.autocast(device_type="cpu", dtype=torch.bfloat16):
+        with torch.autocast(device_type="cpu", dtype=torch.float16):
             onx = to_onnx(
                 model,
                 None,  # args
